+++ packages/plg_task_cstemplateintegrity/services/provider.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * @package     Cstemplateintegrity
+ * @copyright   Copyright (C) 2026 Cybersalt. All rights reserved.
+ * @license     GNU General Public License version 2 or later
+ */
+
+declare(strict_types=1);
+
+defined('_JEXEC') or die;
+
+use Cybersalt\Plugin\Task\Cstemplateintegrity\Extension\Cstemplateintegrity;
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+
+return new class () implements ServiceProviderInterface {
+    public function register(Container $container): void
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $dispatcher = $container->get(DispatcherInterface::class);
+                $plugin     = new Cstemplateintegrity(
+                    $dispatcher,
+                    (array) PluginHelper::getPlugin('task', 'cstemplateintegrity')
+                );
+                $plugin->setApplication(Factory::getApplication());
+
+                return $plugin;
+            }
+        );
+    }
+};